Refactored class names to match Main added inline documentation for Athlete and Athlete specific classes
