--- athlete/src/com/lambdaschool/initial/AthleteService.java
@@ -1,5 +1,0 @@
-package com.lambdaschool.initial;
-
-public interface AthleteService {
-    void displayAthlete(String athlete)
-}
