+++ athlete/src/com/lambdaschool/initial/Main.java
@@ -1,13 +1,15 @@
-package com.lambdaschool.initial;
-
-public class Main
-{
-
-    public static void main(String[] args)
-    {
-        MyApplication myAthlete = new MyApplication();
-        myAthlete.create("Track");
-        myAthlete.create("Hockey");
-        myAthlete.create("Rugby");
-    }
-}
+//package com.lambdaschool.initial;
+//
+//import com.lambdaschool.solution.MyApplication;
+//
+//public class Main
+//{
+//
+//    public static void main(String[] args)
+//    {
+//        MyApplication myAthlete = new MyApplication();
+//        myAthlete.create("Track");
+//        myAthlete.create("Hockey");
+//        myAthlete.create("Rugby");
+//    }
+//}
