--- athlete/src/com/lambdaschool/initial/Processor.java
@@ -1,5 +1,0 @@
-package com.lambdaschool.initial;
-
-public interface Processor {
-    void displayAthlete(String athlete);
-}
