--- athlete/src/com/lambdaschool/initial/TrackServiceImpl.java
@@ -1,8 +1,0 @@
-package com.lambdaschool.initial;
-
-public class TrackServiceImpl implements AthleteService {
-    @Override
-    public void displayAthlete(String athlete) {
-        System.out.println("Track Runner");
-    }
-}
