+++ athlete/src/com/lambdaschool/solution/AthleteCreation.java
@@ -1,0 +1,5 @@
+package com.lambdaschool.solution;
+
+public interface AthleteCreation {
+    void displayAthlete();
+}
