+++ athlete/src/com/lambdaschool/solution/AthleteCreationInjector.java
@@ -1,0 +1,5 @@
+package com.lambdaschool.solution;
+
+public interface AthleteCreationInjector {
+    Processor getProcess();
+}
