+++ athlete/src/com/lambdaschool/solution/Processor.java
@@ -1,0 +1,5 @@
+package com.lambdaschool.solution;
+
+public interface Processor {
+    void displayAthlete();
+}
