+++ athlete/src/com/lambdaschool/solution/Shoes.java
@@ -1,4 +1,4 @@
-package com.lambdaschool.initial;
+package com.lambdaschool.solution;
 
 public class Shoes
 {
