+++ athlete/src/com/lambdaschool/solution/TrackAthleteCreationImpl.java
@@ -1,0 +1,9 @@
+package com.lambdaschool.solution;
+
+public class TrackAthleteCreationImpl implements AthleteCreation {
+    // The athlete specific Implementor implements the generic interface AthleteCreation requiring it to use displayAthlete method declared within interface.
+    @Override
+    public void displayAthlete() {
+        System.out.println("Track Runner");
+    }
+}
